requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-groq
langchain-community
qdrant-client
pymongo
bcrypt
python-dotenv
sentence-transformers
pypdf